+++ tests/TestCase/Controller/Component/UploaderComponentTest.php
@@ -33,6 +33,12 @@ use MeTools\Controller\Component\UploaderComponent as BaseUploaderComponent;
  */
 class UploaderComponent extends BaseUploaderComponent
 {
+    protected function move_uploaded_file($filename, $destination)
+    {
+        //@codingStandardsIgnoreLine
+        return @rename($filename, $destination);
+    }
+
     public function getFile()
     {
         if (!isset($this->file)) {
@@ -45,6 +51,11 @@ class UploaderComponent extends BaseUploaderComponent
     public function findTargetFilename($target)
     {
         return $this->_findTargetFilename($target);
+    }
+
+    public function resetError()
+    {
+        unset($this->error);
     }
 
     public function setError($error)
@@ -62,6 +73,25 @@ class UploaderComponentTest extends TestCase
      * @var \UploaderComponent
      */
     protected $Uploader;
+
+    /**
+     * Internal method to create a file and get a valid array for upload
+     * @return array
+     */
+    protected function _createFile()
+    {
+        //Creates a file and writes some content
+        $file = tempnam(TMP, 'php_upload_');
+        file_put_contents($file, 'string');
+
+        return [
+            'name' => basename($file),
+            'type' => mime_content_type($file),
+            'tmp_name' => $file,
+            'error' => UPLOAD_ERR_OK,
+            'size' => filesize($file),
+        ];
+    }
 
     /**
      * Setup the test case, backup the static object values so they can be
@@ -85,6 +115,11 @@ class UploaderComponentTest extends TestCase
     public function tearDown()
     {
         parent::tearDown();
+
+        //Deletes all files
+        foreach (glob(UPLOADS . '*') as $file) {
+            unlink($file);
+        }
 
         unset($this->Uploader);
     }
@@ -111,9 +146,9 @@ class UploaderComponentTest extends TestCase
      */
     public function testFindTargetFilename()
     {
-        $file1 = TMP . 'target.txt';
-        $file2 = TMP . 'target_1.txt';
-        $file3 = TMP . 'target_2.txt';
+        $file1 = UPLOADS . 'target.txt';
+        $file2 = UPLOADS . 'target_1.txt';
+        $file3 = UPLOADS . 'target_2.txt';
 
         $this->assertEquals($file1, $this->Uploader->findTargetFilename($file1));
 
@@ -125,28 +160,157 @@ class UploaderComponentTest extends TestCase
         file_put_contents($file2, null);
         $this->assertEquals($file3, $this->Uploader->findTargetFilename($file1));
 
-        unlink($file1);
-        unlink($file2);
+        //Files without extension
+        $file1 = UPLOADS . 'target';
+        $file2 = UPLOADS . 'target_1';
+
+        $this->assertEquals($file1, $this->Uploader->findTargetFilename($file1));
+
+        //Creates the first file
+        file_put_contents($file1, null);
+        $this->assertEquals($file2, $this->Uploader->findTargetFilename($file1));
     }
 
     /**
      * Tests for `set()` method
      * @test
+     * @uses _createFile()
      */
     public function testSet()
     {
-        $result = $this->Uploader->set(['error' => UPLOAD_ERR_OK]);
-        $this->assertEquals('MeTools\Test\TestCase\Controller\Component\UploaderComponent', get_class($result));
-        $this->assertNotFalse($this->Uploader->getFile());
+        $file = $this->_createFile();
 
-        $this->Uploader->set(['error' => UPLOAD_ERR_INI_SIZE]);
-        $error = $this->Uploader->error();
-        $this->assertEquals('The uploaded file exceeds the maximum size that was specified in php.ini', $error);
+        $this->Uploader->set($file);
+        $this->assertEmpty($this->Uploader->error());
+        $this->assertEquals('stdClass', get_class($this->Uploader->getFile()));
+        $this->assertEquals(
+            ['name', 'type', 'tmp_name', 'error', 'size'],
+            array_keys((array)$this->Uploader->getFile())
+        );
+
+        $this->Uploader->set(array_merge($file, ['error' => UPLOAD_ERR_INI_SIZE]));
+        $this->assertNotEmpty($this->Uploader->error());
         $this->assertFalse($this->Uploader->getFile());
 
-        $this->Uploader->set(['error' => 'noExistingErrorCode']);
-        $error = $this->Uploader->error();
-        $this->assertEquals('Unknown upload error', $error);
+        $this->Uploader->set(array_merge($file, ['error' => 'noExistingErrorCode']));
+        $this->assertEquals('Unknown upload error', $this->Uploader->error());
         $this->assertFalse($this->Uploader->getFile());
+    }
+
+    /**
+     * Test for `mimetype()` method
+     * @test
+     * @uses _createFile()
+     */
+    public function testMimetype()
+    {
+        $file = $this->_createFile();
+        $this->Uploader->set($file);
+
+        $this->Uploader->mimetype('text/plain');
+        $this->assertEmpty($this->Uploader->error());
+
+        $this->Uploader->mimetype('text');
+        $this->assertEmpty($this->Uploader->error());
+
+        $this->Uploader->mimetype(['text/plain', 'image/gif']);
+        $this->assertEmpty($this->Uploader->error());
+
+        $this->Uploader->mimetype('image/gif');
+        $this->assertEquals('The mimetype image/gif is not accepted', $this->Uploader->error());
+
+        //Resets error
+        $this->Uploader->resetError();
+
+        $this->Uploader->mimetype(['image/gif', 'image/png']);
+        $this->assertEquals('The mimetype image/gif, image/png is not accepted', $this->Uploader->error());
+    }
+
+    /**
+     * Test for `mimetype()` method, with no file
+     * @expectedException Cake\Network\Exception\InternalErrorException
+     * @expectedExceptionMessage There are no uploaded file information
+     * @test
+     */
+    public function testMimetypeNoFile()
+    {
+        $this->Uploader->mimetype('text/plain');
+    }
+
+    /**
+     * Test for `save()` method
+     * @test
+     * @uses _createFile()
+     */
+    public function testSave()
+    {
+        $file = $this->_createFile();
+        $this->assertFileExists($file['tmp_name']);
+        $this->Uploader->set($file);
+
+        $result = $this->Uploader->save(UPLOADS);
+        $this->assertFalse($this->Uploader->error());
+        $this->assertFileExists($result);
+        $this->assertFileNotExists($file['tmp_name']);
+    }
+
+    /**
+     * Test for `save()` method, with a not writable directory
+     * @test
+     * @uses _createFile()
+     */
+    public function testSaveNoWritableDir()
+    {
+        //Creates a non-writable directory
+        //@codingStandardsIgnoreLine
+        @mkdir(TMP . 'noWritableDir', 0444);
+
+        $file = $this->_createFile();
+        $this->Uploader->set($file);
+
+        $this->assertFalse($this->Uploader->save(TMP . 'noWritableDir'));
+        $this->assertEquals('The file was not successfully moved to the target directory', $this->Uploader->error());
+    }
+
+    /**
+     * Test for `save()` method, using a no existing directory
+     * @expectedException Cake\Network\Exception\InternalErrorException
+     * @expectedExceptionMessage Invalid or no existing directory /tmp/uploads/noExistingDir
+     * @test
+     * @uses _createFile()
+     */
+    public function testSaveNoExistingDir()
+    {
+        $file = $this->_createFile();
+        $this->Uploader->set($file)->save(UPLOADS . 'noExistingDir');
+    }
+
+    /**
+     * Test for `save()` method, with no file
+     * @expectedException Cake\Network\Exception\InternalErrorException
+     * @expectedExceptionMessage There are no uploaded file information
+     * @test
+     */
+    public function testSaveNoFile()
+    {
+        $this->Uploader->save(null);
+    }
+
+    /**
+     * Test for `save()` method, with an error
+     * @test
+     * @uses _createFile()
+     */
+    public function testSaveWithError()
+    {
+        $file = $this->_createFile();
+        $this->Uploader->set($file);
+
+        //Sets an error
+        $error = 'error before save';
+        $this->Uploader->setError($error);
+
+        $this->assertFalse($this->Uploader->save(UPLOADS));
+        $this->assertEquals($error, $this->Uploader->error());
     }
 }