+++ tests/TestCase/View/Helper/BBCodeHelperTest.php
@@ -33,6 +33,16 @@ use MeTools\View\Helper\HtmlHelper;
 class BBCodeHelperTest extends TestCase
 {
     /**
+     * @var \MeTools\View\Helper\BBCodeHelper
+     */
+    protected $BBCode;
+
+    /**
+     * @var \MeTools\View\Helper\HtmlHelper
+     */
+    protected $Html;
+
+    /**
      * Setup the test case, backup the static object values so they can be
      * restored. Specifically backs up the contents of Configure and paths in
      *  App if they have not already been backed up